+++ src/tree.c
@@ -355,7 +355,7 @@ size_t git_tree_entrycount(const git_tree *tree)
 unsigned int git_treebuilder_entrycount(git_treebuilder *bld)
 {
 	assert(bld);
-	return bld->entries.length;
+	return (int)bld->entries.length;
 }
 
 static int tree_error(const char *str, const char *path)
